Safari-611.1.10/LayoutTests/fast/mediastream/MediaStreamTrack-kind-expected.txt

Test MediaStreamTrack kind

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaStream.getVideoTracks().length is 1
PASS mediaStream.getAudioTracks().length is 1
PASS mediaStream.getAudioTracks()[0].kind is "audio"
PASS mediaStream.getVideoTracks()[0].kind is "video"
PASS successfullyParsed is true

TEST COMPLETE
